Fix the tests failing after merge, remove duplicated tests
